Serialization
Celery is implemented, but because of the intense use of files and folders it is not sufficient.

new methods:
- ymap_webtool/tasks.py.is_too_old(file, age_min)
   checks if a file is older then X minutes

updated methods:
-  ymap_webtool/view.py.index(request)
-  ymap_webtool/view.py.get_file(request)
-  ymap_webtool/view.py.download_result(request)

new templates:
- server_is_busy.html

Serialization is now solved in a way, that once a job is submitted, a file (job_running) is created, and when the user downloads the results, this file is deleted.
As long as job_running exists, the server prevents another submission.
If the user clicks away without downloading the results, the job_running file is deleted upon request for the index page, IF job_running is more than 15 minutes old.